+++ src/commands/taskkill.js
@@ -1,8 +1,6 @@
 const chalk = require("chalk");
 const psList = require("ps-list");
 
-const { GLOBAL_NAME } = require("../variables/constants");
-
 const _promptForYN = require("../functions/promptForYN");
 const _fatalError = require("../functions/fatalError");
 
@@ -10,32 +8,7 @@ const Errors = require("../classes/Errors");
 const Checks = require("../classes/Checks");
 const InfoMessages = require("../classes/InfoMessages");
 const Verbose = require("../classes/Verbose");
-
-/**
- * Checks if the user wants to kill the BubbleOS process,
- * and gives a message if so.
- *
- * @param {string | number} pid The PID that the user entered.
- * @param {boolean} killSelf Whether or not the user allowed BubbleOS to kill itself.
- * @returns `true` if the user attempted to kill BubbleOS, `false` otherwise.
- */
-const _killSelfMsg = (pid, killSelf) => {
-  // If the PID is equal to BubbleOS' PID and the user did not give permission to kill itself
-  if (Number(pid) === process.pid && !killSelf) {
-    Verbose.custom("Attempted to kill BubbleOS process...");
-    console.log(
-      chalk.yellow(
-        `You cannot kill the ${GLOBAL_NAME} process. To exit ${GLOBAL_NAME}, run the '${chalk.italic(
-          "exit"
-        )}' command.\nProcess aborted.\n`
-      )
-    );
-
-    return true;
-  }
-
-  return false;
-};
+const PathUtil = require("../classes/PathUtil");
 
 /**
  * Kill a process on the device from the BubbleOS
@@ -47,9 +20,6 @@ const _killSelfMsg = (pid, killSelf) => {
  * prompt.
  * - `-s`: Silence all success messages, excluding
  * error messages and the confirmation prompt.
- * - `--kill-self`: Allow killing the BubbleOS process.
- * Use this flag at your own risk, as it can remove any
- * unsaved data in BubbleOS!
  *
  * @param {string | number} processName The PID or process name to kill.
  * @param {...string} args Arguments to change the behavior of `taskkill` (listed above).
@@ -62,7 +32,9 @@ const taskkill = async (processName, ...args) => {
     Verbose.initArgs();
     const confirm = !args?.includes("-y");
     const silent = args?.includes("-s");
-    const killSelf = args?.includes("--kill-self");
+
+    Verbose.parseQuotes();
+    processName = PathUtil.parseQuotes([processName, ...args])[0];
 
     if (processChk.paramUndefined()) {
       Verbose.chkEmpty();
@@ -80,8 +52,6 @@ const taskkill = async (processName, ...args) => {
     }
 
     if (!isNaN(Number(processName))) {
-      if (_killSelfMsg(processName, killSelf)) return;
-
       Verbose.custom("Attempting to kill process...");
       process.kill(Number(processName));
 
@@ -93,10 +63,14 @@ const taskkill = async (processName, ...args) => {
       const processes = await psList();
       const result = processes.find((obj) => obj.name === processName);
 
+      if (!result) {
+        Verbose.custom(`The process '${processName}' was detected to not exist.`);
+        Errors.doesNotExist("process", processName);
+        return;
+      }
+
       processes.forEach((obj) => {
         if (obj.name === processName) {
-          if (_killSelfMsg(result.pid, killSelf)) return;
-
           Verbose.custom("Attempting to kill process...");
           try {
             process.kill(Number(result.pid));